requirements.txt
python-dotenv
streamlit
Pillow
openai